+++ Jint/Native/Function/BindFunctionInstance.cs
@@ -6,7 +6,8 @@ namespace Jint.Native.Function
 {
     public sealed class BindFunctionInstance : FunctionInstance, IConstructor
     {
-        public BindFunctionInstance(Engine engine) : base(engine, null)
+        public BindFunctionInstance(Engine engine) 
+            : base(engine, name: null, thisMode: FunctionThisMode.Strict)
         {
         }
